requirements.txt
anyio==3.6.1
argon2-cffi==21.3.0
argon2-cffi-bindings==21.2.0
asgiref==3.5.2
asttokens==2.0.5
attrs==21.4.0
Babel==2.10.3
backcall==0.2.0
backports.zoneinfo==0.2.1
beautifulsoup4==4.11.1
bleach==5.0.1
certifi==2022.6.15
cffi==1.15.0
charset-normalizer==2.0.12
cmaths==0.1.0
colorama==0.4.5
countryinfo==0.1.2
debugpy==1.6.0
decorator==5.1.1
defusedxml==0.7.1
distlib==0.3.5
Django==4.0.6
django-debug-toolbar==3.5.0
django-extensions==3.2.0
django-filter==22.1
django-phonenumber-field==6.3.0
djangorestframework==3.13.1
emoji==1.7.0
entrypoints==0.4
executing==0.8.3
fastjsonschema==2.15.3
filelock==3.7.1
idna==3.3
importlib-metadata==4.12.0
importlib-resources==5.8.0
ipykernel==6.15.0
ipython==8.4.0
ipython-genutils==0.2.0
jedi==0.18.1
Jinja2==3.1.2
json5==0.9.8
jsonschema==4.6.1
jupyter-client==7.3.4
jupyter-core==4.10.0
jupyter-server==1.18.0
jupyterlab==3.4.3
jupyterlab-pygments==0.2.2
jupyterlab-server==2.14.0
MarkupSafe==2.1.1
matplotlib-inline==0.1.3
mistune==0.8.4
nbclassic==0.3.7
nbclient==0.6.4
nbconvert==6.5.0
nbformat==5.4.0
nest-asyncio==1.5.5
notebook==6.4.12
notebook-shim==0.1.0
null==0.6.1
numpy==1.23.0
packaging==21.3
pandas==1.4.2
pandocfilters==1.5.0
parso==0.8.3
phonenumbers==8.12.53
pickleshare==0.7.5
platformdirs==2.5.2
prometheus-client==0.14.1
prompt-toolkit==3.0.30
psutil==5.9.1
psycopg2==2.9.3
pure-eval==0.2.2
pycparser==2.21
Pygments==2.12.0
PyInputPlus==0.2.12
pyparsing==3.0.9
pyrsistent==0.18.1
PySimpleValidate==0.2.12
python-dateutil==2.8.2
pytz==2022.1
pywin32==304
pywinpty==2.0.5
pyzmq==23.2.0
requests==2.28.0
scipy==1.8.1
Send2Trash==1.8.0
six==1.16.0
sniffio==1.2.0
soupsieve==2.3.2.post1
sqlparse==0.4.2
stack-data==0.3.0
stdiomask==0.0.6
tabulate==0.8.10
terminado==0.15.0
tinycss2==1.1.1
tornado==6.1
traitlets==5.3.0
tzdata==2022.1
urllib3==1.26.9
virtualenv==20.16.2
wcwidth==0.2.5
webencodings==0.5.1
websocket-client==1.3.3
zipp==3.8.0
ZMATH==1.0
